requirements_pip.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.12.0
joblib==1.1.0
kiwisolver==1.4.4
liftoff @ git+https://github.com/tudor-berariu/liftoff.git@4a0734006fb15c6da5013d437263161d3facf3d8
Markdown==3.3.6
matplotlib==3.7.1
multidict==6.0.2
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandas==2.0.0
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.8
pyperclip==1.8.2
python-dateutil==2.8.2
pytorch-lightning==1.5.6
pytz==2023.3
PyYAML==6.0
random-fourier-features-pytorch==1.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smmap==5.0.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.5.4
torch==1.11.0
torchmetrics==0.8.1
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2023.3
urllib3==1.26.9
Werkzeug==2.1.2
xgboost==1.7.5
yarl==1.7.2
zipp==3.8.0
